pyproject.toml
[project]
name = "ducku"
version = "1.1.0"
description = "Documentation analysis and code quality tool for multi-language projects"
readme = "README.md"
license = {text = "Ducku Source Available License (DSAL)"}
authors = [
    {name = "Roman Krivtsov", email = ""}
]
maintainers = [
    {name = "Roman Krivtsov", email = ""}
]
keywords = ["documentation", "code-analysis", "static-analysis", "dead-code", "linting"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Documentation",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Utilities",
]
requires-python = ">=3.10"
dependencies = [
    "anytree>=2.13.0",
    "codespell>=2.4.1",
    "jsonschema>=4.25.1",
    "markdown-it-py>=4.0.0",
    "marko>=2.2.0",
    "pyyaml",
    "rapidfuzz>=3.14.0",
    "tree-sitter>=0.23.0",
    "tree-sitter-python",
    "tree-sitter-javascript",
    "tree-sitter-typescript",
    "tree-sitter-java",
    "tree-sitter-php",
    "tree-sitter-rust",
    "tree-sitter-ruby",
    "tree-sitter-go",
    "pydantic>=2.12.5",
]

[project.urls]
Homepage = "https://github.com/duckuio/ducku_cli"
Repository = "https://github.com/duckuio/ducku_cli"
Issues = "https://github.com/duckuio/ducku_cli/issues"
Documentation = "https://github.com/duckuio/ducku_cli#readme"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "twine>=6.2.0",
]
build = [
    "build>=1.0.0",
    "twine>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "bin"]

[tool.pytest.ini_options]
pythonpath = [
    "."
]
testpaths = [
    "test"
]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
[project.scripts]
ducku = "bin.cli:main"
